confirm reset

+++ momentum/js/reset.js
@@ -1,6 +1,9 @@
 const resetBtn = document.querySelector(".reset-btn");
 
 function handleReset() {
-    localStorage.clear();
+    let reallyReset = confirm("정말로 리셋하시겠습니까?");
+    if (reallyReset) {
+        localStorage.clear();
+    }
 }
 resetBtn.addEventListener("click", handleReset);
